Get the tests working again.
If we have two modules with two main.swift during testing the linker
will try to link both of them into a single executable and get a
conflict. So all the modules from the various days need to be split
out into their own packages and the main executables left untested.